tax-review-app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
anthropic>=0.39.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
pymupdf>=1.23.0
openpyxl>=3.1.2
pandas>=2.1.0
Pillow>=10.1.0
python-dateutil>=2.8.2
aiofiles>=23.2.1
